ml_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
pydantic==2.7.0
openai==1.16.2
